Filter clusters by mapping items
