+++ app/javascript/react/screens/App/Overview/screens/PlanWizard/PlanWizardSelectors.js
@@ -1,4 +1,5 @@
 import { getMappingType } from '../../../Mappings/components/InfrastructureMappingsList/helpers';
+import { TRANSFORMATION_MAPPING_ITEM_DESTINATION_TYPES } from '../../../../../../common/constants';
 
 export const findEditingPlan = (transformationPlans, editingPlanId) =>
   editingPlanId && transformationPlans.find(plan => plan.id === editingPlanId);
@@ -45,11 +46,23 @@ export const getWarmMigrationCompatibility = ({
   const isEveryVmCompatible = vms.every(vm => vm.warm_migration_compatible);
 
   const selectedMapping = transformationMappings.find(map => map.id === generalStepValues.infrastructure_mapping);
-  console.log('warm migration compat?', { selectedMapping, targetProviderType, targetClusters, conversionHosts });
+  const targetClusterType = TRANSFORMATION_MAPPING_ITEM_DESTINATION_TYPES[targetProviderType].cluster;
+  const targetClustersInSelectedMapping = selectedMapping.transformation_mapping_items
+    .filter(mappingItem => mappingItem.destination_type === targetClusterType)
+    .map(mappingItem => targetClusters.find(cluster => cluster.id === mappingItem.destination_id));
+
+  console.log('warm migration compat?', {
+    selectedMapping,
+    targetProviderType,
+    targetClusters,
+    conversionHosts,
+    targetClustersInSelectedMapping,
+    vms
+  });
 
   // * Figure out a list of the target clusters associated with the selected VMs
   //   - For each mapping item of the corresponding destination_type (EmsCluster or CloudTenant) find the matching cluster by destination_id (use TRANSFORMATION_MAPPING_ITEM_DESTINATION_TYPES)
-  //   - Filter by associated VM? how does this work?
+  //   - Filter by associated VM? how does this work? <----- ???
   // * For every cluster (.every()):
   //   - Figure out the EMS id of each target cluster (use ems_id of the loaded cluster object)
   //   - Find conversion hosts whose resource have that EMS id
